backend/requirements.txt
Flask
PyMuPDF
python-dotenv
google-generativeai
Flask-CORS 